Picking Tutorial Complete
So this tutorial was fairly easy. It is still kinda awkward as the cursor seems to default to a loading symbol due to the running nature of the program. Googling has shown no quick solution so this issue has been left for the future since usually you would not see the cursor